First commit
I worked for a couple of hours on this calculator, so a lot of steps are missing. For now, it can do all basics operation (addition, substraction, multiplication and division) and calculate the remainder (modulo). There can be theoretically an infinity of operands, and the current operations take into account the order of operations. 

A lot of optimization was done, and there are some subtleties (when their is nothing after the dot, the calculator will show the result without a dot. For example, if the result is 9.0, it will display "9", when normally a java program would have display "9.0", since the variable "result" is a double).