Recursion
The most important change: the computation of operations is now set in a recursive method, which will probably make the implementation of parentheses easier. 
Some other changes: the new command help to give users more explanations (and a more detailed welcome message), the verification that the input equals 'q' (and not that it contains it), the discovery of A NEW BUG (negative numbers are not only a problem as operands, but as results too), et voila.